backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.4
bcrypt==4.1.2
pyjwt==2.8.0
python-multipart==0.0.9